pyproject.toml
[project]
name = "cal-helper"
version = "0.1.0"
description = "Scrape events from chalmers.it"
license = { file = 'LICENSE' }
authors = [{ name = 'Oli', email = 'oli@olillin.com' }]
maintainers = [{ name = 'Oli', email = 'oli@olillin.com' }]

requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]

[project.urls]
"Bug Reports" = "https://github.com/olillin/cal-helper/issues"
"Source" = "https://github.com/olillin/cal-helper"

[project.scripts]
cal-helper = 'cal_helper:main'

[build-system]
requires = ['uv_build>=0.9.0,<0.10.0']
build-backend = 'uv_build'
